fix: Move TanStack Table column definitions to client components
Fixes React 19 error: "Functions cannot be passed directly to Client Components"

Changes:
- Created separate columns.tsx files with "use client" directive
- Moved column definitions for tenants, properties, and maintenance pages
- Follows Next.js 15 + React 19 best practices (October 2025)
- Implements industry-standard Shadcn/UI data-table pattern

Pattern:
- columns.tsx: Client component with column definitions
- page.tsx: Server component for data fetching

This resolves the Server/Client Component boundary violation that was
preventing pages from rendering correctly.
