+++ apps/frontend/src/app/(protected)/manage/tenants/page.tsx
@@ -2,16 +2,13 @@ import type { Metadata } from 'next'
 import { requireSession } from '@/lib/server-auth'
 import { createLogger } from '@repo/shared/lib/frontend-logger'
 import { tenantsApi } from '@/lib/api-client'
-import { Badge } from '@/components/ui/badge'
 import { Button } from '@/components/ui/button'
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
-import { DataTable } from '@/components/ui/data-table'
-import { AlertDialog, AlertDialogAction, AlertDialogCancel, AlertDialogContent, AlertDialogDescription, AlertDialogFooter, AlertDialogHeader, AlertDialogTitle, AlertDialogTrigger } from '@/components/ui/alert-dialog'
-import { Trash2, Plus } from 'lucide-react'
+import { Card, CardDescription, CardHeader, CardTitle } from '@/components/ui/card'
+import { Plus } from 'lucide-react'
 import Link from 'next/link'
-import { ColumnDef } from '@tanstack/react-table'
 import type { TenantStats, TenantWithLeaseInfo } from '@repo/shared/types/core'
 import { TenantsTableClient } from './tenants-table.client'
+import { columns } from './columns'
 
 export const metadata: Metadata = {
 	title: 'Tenants | TenantFlow',
@@ -44,91 +41,6 @@ export default async function TenantsPage() {
 			error: err instanceof Error ? err.message : String(err)
 		})
 	}
-
-	// ✅ Inline columns - NO wrapper file
-	const columns: ColumnDef<TenantWithLeaseInfo>[] = [
-		{
-			accessorKey: 'name',
-			header: 'Tenant',
-			cell: ({ row }) => {
-				const tenant = row.original
-				return (
-					<Link href={`/manage/tenants/${tenant.id}`} className="hover:underline">
-						<div className="flex flex-col">
-							<span className="font-medium">{tenant.name}</span>
-							<span className="text-sm text-muted-foreground">{tenant.email}</span>
-						</div>
-					</Link>
-				)
-			}
-		},
-		{
-			accessorKey: 'property',
-			header: 'Property',
-			cell: ({ row }) => {
-				const tenant = row.original
-				return tenant.property?.name ? (
-					<div className="flex flex-col">
-						<span>{tenant.property.name}</span>
-						<span className="text-sm text-muted-foreground">
-							{tenant.property.city}, {tenant.property.state}
-						</span>
-					</div>
-				) : (
-					<span className="text-muted-foreground">No property</span>
-				)
-			}
-		},
-		{
-			accessorKey: 'paymentStatus',
-			header: 'Status',
-			cell: ({ row }) => {
-				const tenant = row.original
-				return (
-					<div className="flex flex-col gap-1">
-						<Badge
-							variant={
-								tenant.paymentStatus === 'Overdue'
-									? 'destructive'
-									: tenant.paymentStatus === 'Current'
-										? 'secondary'
-										: 'outline'
-							}
-						>
-							{tenant.paymentStatus}
-						</Badge>
-						<Badge variant="outline">{tenant.leaseStatus}</Badge>
-					</div>
-				)
-			}
-		},
-		{
-			accessorKey: 'currentLease',
-			header: 'Lease',
-			cell: ({ row }) => {
-				const tenant = row.original
-				return tenant.currentLease ? (
-					<div className="flex flex-col">
-						<span className="text-sm">#{tenant.currentLease.id.slice(0, 8)}</span>
-						<span className="text-sm text-muted-foreground">
-							{tenant.leaseStart ? new Date(tenant.leaseStart).toLocaleDateString() : 'Start TBD'} -{' '}
-							{tenant.leaseEnd ? new Date(tenant.leaseEnd).toLocaleDateString() : 'End TBD'}
-						</span>
-					</div>
-				) : (
-					<span className="text-muted-foreground">No active lease</span>
-				)
-			}
-		},
-		{
-			accessorKey: 'monthlyRent',
-			header: 'Monthly Rent',
-			cell: ({ row }) => {
-				const rent = row.getValue('monthlyRent') as number | null
-				return rent ? new Intl.NumberFormat('en-US', { style: 'currency', currency: 'USD' }).format(rent) : '-'
-			}
-		}
-	]
 
 	return (
 		<div className="space-y-8">
